Added "No PDFs found"
Added a message that will be displayed if no PDF's were found.  The message is a simple default li tag containing a div that will be removed if the count is greater than 0.

I think a search field will be a nice feature at some point.  If it's added, it may be better to make the visibility hidden rather tan remove.  That way, it can be made visible if the search yields no results.